requirements.txt
# Music Linker Requirements
# Python dependencies for the IMDb soundtrack to YouTube linker

# Core dependencies
pydantic>=2.0.0
google-api-python-client>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Optional but recommended
requests>=2.31.0
